project/app/function-source/requirements.txt
Pillow
tensorflow
numpy
zipfile36
functions-framework==3.*
wget
requests
google.cloud.storage
